llm_applications_using_langchain/Q&A_chatbot_using_langchain/requirements.txt
langchain-openai
langchain
python-dotenv 
langchain_community 
langchain_core 
streamlit 